requirements.txt
Django>=4.2,<5.0
django-tenants>=3.6,<4.0
djangorestframework>=3.14,<4.0
django-otp>=1.3,<2.0
qrcode[pil]>=7.4
pyotp>=2.9
argon2-cffi>=23.1
django-ratelimit>=4.1,<5.0
gunicorn>=21.2
psycopg2-binary>=2.9
python-dotenv>=1.0
whitenoise>=6.5
